backend/requirements.txt

django==5.2.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==0.19
docxtpl==0.20.0
